# contracts/admin-api.openapi.yaml
openapi: 3.1.0
info:
  title: AI Control Layer – Admin API
  description: Management and reporting API consumed by the admin panel.
  version: 1.0.0
  x-gateway-version: 0.1.0
paths:
...
  /admin/v1/approvals:
    get:
      tags:
      - approvals
      summary: List Approvals
      operationId: listApprovals
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ApprovalStatus'
          - type: 'null'
          default: pending
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Approval'
                title: Response Listapprovals
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/approvals/{approval_id}:
    get:
      tags:
      - approvals
      summary: Get Approval
      operationId: getApprovalAdmin
      security:
      - HTTPBearer: []
      parameters:
      - name: approval_id
        in: path
        required: true
        schema:
          type: string
          title: Approval Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/approvals/{approval_id}/decision:
    post:
      tags:
      - approvals
      summary: Decide Approval
      description: Approve (optionally with time-boxed elevation) or deny. Audited.
      operationId: decideApproval
      security:
      - HTTPBearer: []
      parameters:
      - name: approval_id
        in: path
        required: true
        schema:
          type: string
          title: Approval Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/budgets:
    get:
      tags:
      - budgets
      summary: Budgets
      operationId: getBudgets
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetTree'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/budgets/breakers:
    get:
      tags:
      - budgets
      summary: Breakers
      operationId: listBreakers
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BreakerState'
                type: array
                title: Response Listbreakers
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/budgets/breakers/{breaker_id}/reset:
    post:
      tags:
      - budgets
      summary: Reset Breaker
      operationId: resetBreaker
      security:
      - HTTPBearer: []
      parameters:
      - name: breaker_id
        in: path
        required: true
        schema:
          type: string
          title: Breaker Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakerState'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/connectors:
    get:
      tags:
      - models
      summary: Connectors
      operationId: listConnectors
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ConnectorStatus'
                type: array
                title: Response Listconnectors
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/connectors/{connector_id}/kill-switch:
    post:
      tags:
      - models
      summary: Kill Switch
      operationId: setKillSwitch
      security:
      - HTTPBearer: []
      parameters:
      - name: connector_id
        in: path
        required: true
        schema:
          type: string
          title: Connector Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KillSwitchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorStatus'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/models:
    get:
      tags:
      - models
      summary: Models
      operationId: listModelsAdmin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModelInfo'
                type: array
                title: Response Listmodelsadmin
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/mcp/servers:
    get:
      tags:
      - mcp
      summary: Mcp Servers
      operationId: listMcpServers
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/McpServerInfo'
                type: array
                title: Response Listmcpservers
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/mcp/tools:
    get:
      tags:
      - mcp
      summary: Mcp Tools
      operationId: listMcpTools
      security:
      - HTTPBearer: []
      parameters:
      - name: server
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Server
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/McpToolInfo'
                title: Response Listmcptools
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/mcp/tools/{tool_id}/approve:
    post:
      tags:
      - mcp
      summary: Approve Tool
      description: Re-pin a drifted/pending tool to its current hash.
      operationId: approveMcpTool
      security:
      - HTTPBearer: []
      parameters:
      - name: tool_id
        in: path
        required: true
        schema:
          type: string
          title: Tool Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpToolApprovalRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpToolInfo'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/mcp/tools/{tool_id}/quarantine:
    post:
      tags:
      - mcp
      summary: Quarantine Tool
      operationId: quarantineMcpTool
      security:
      - HTTPBearer: []
      parameters:
      - name: tool_id
        in: path
        required: true
        schema:
          type: string
          title: Tool Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpToolApprovalRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpToolInfo'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/feed:
    get:
      tags:
      - feed
      summary: Feed Status
      operationId: getFeedStatus
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedStatus'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/feed/sync:
    post:
      tags:
      - feed
      summary: Feed Sync
      operationId: syncFeed
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedStatus'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/artifacts/scan:
    post:
      tags:
      - artifacts
      summary: Scan Artifact
      operationId: scanArtifact
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_scanArtifact'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactScanResult'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /admin/v1/artifacts:
    get:
      tags:
      - artifacts
      summary: List Artifacts
      operationId: listArtifacts
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ArtifactScanResult'
                type: array
                title: Response Listartifacts
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/insights/clusters:
    get:
      tags:
      - insights
      summary: Insights
      operationId: listInsights
      security:
      - HTTPBearer: []
      parameters:
      - name: group
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InsightCluster'
                title: Response Listinsights
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/insights/clusters/{cluster_id}/publish:
    post:
      tags:
      - insights
      summary: Publish
      operationId: publishSkill
      security:
      - HTTPBearer: []
      parameters:
      - name: cluster_id
        in: path
        required: true
        schema:
          type: string
          title: Cluster Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishSkillRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsightCluster'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    Approval:
      properties:
        id:
          type: string
          title: Id
        status:
          $ref: '#/components/schemas/ApprovalStatus'
        approver_scope:
          type: string
          enum:
          - user
          - admin
          title: Approver Scope
        created_at:
          type: string
          format: date-time
          title: Created At
        expires_at:
          type: string
          format: date-time
          title: Expires At
        requested_by:
          type: string
          title: Requested By
        session_id:
          type: string
          title: Session Id
        trace_id:
          type: string
          title: Trace Id
        tool:
          anyOf:
          - type: string
          - type: 'null'
          title: Tool
        server:
          anyOf:
          - type: string
          - type: 'null'
          title: Server
        arguments_preview:
          type: string
          title: Arguments Preview
          description: Redacted preview (command, diff, recipients).
          default: ''
        reason:
          type: string
          title: Reason
          default: ''
        rule_ids:
          items:
            type: string
            pattern: ^[A-Z][A-Z0-9]*(-[A-Z0-9_.]+)+$
            description: Rule identifier (control id, or control id + sub-rule, or signature id).
          type: array
          title: Rule Ids
        risk_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Risk Score
          default: 0.0
        decided_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Decided By
        decided_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Decided At
        elevation:
          anyOf:
          - $ref: '#/components/schemas/Elevation'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - id
      - status
      - approver_scope
      - created_at
      - expires_at
      - requested_by
      - session_id
      - trace_id
      title: Approval
    ApprovalDecisionRequest:
      properties:
        decision:
          type: string
          enum:
          - approve
          - deny
          title: Decision
        elevation_minutes:
          anyOf:
          - type: integer
            maximum: 240.0
            minimum: 0.0
          - type: 'null'
          title: Elevation Minutes
        note:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - decision
      title: ApprovalDecisionRequest
...
    ApprovalStatus:
      type: string
      enum:
      - pending
      - approved
      - denied
      - expired
      title: ApprovalStatus
...
    ArtifactScanResult:
      properties:
        id:
          type: string
          title: Id
        filename:
          type: string
          title: Filename
        sha256:
          type: string
          title: Sha256
        size:
          type: integer
          title: Size
        format_detected:
          type: string
          title: Format Detected
        verdict:
          type: string
          enum:
          - safe
          - malicious
          - blocked_format
          - suspicious
          title: Verdict
        findings:
          items:
            $ref: '#/components/schemas/ArtifactFinding'
          type: array
          title: Findings
        scanned_at:
          type: string
          format: date-time
          title: Scanned At
      additionalProperties: false
      type: object
      required:
      - id
      - filename
      - sha256
      - size
      - format_detected
      - verdict
      - scanned_at
      title: ArtifactScanResult
...
    Body_scanArtifact:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_scanArtifact
...
    BreakerState:
      properties:
        id:
          type: string
          title: Id
        state:
          type: string
          enum:
          - closed
          - open
          - half_open
          title: State
        opened_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Opened At
        cooldown_until:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cooldown Until
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      additionalProperties: false
      type: object
      required:
      - id
      - state
      title: BreakerState
...
    BudgetTree:
      properties:
        generated_at:
          type: string
          format: date-time
          title: Generated At
        nodes:
          items:
            $ref: '#/components/schemas/BudgetNode'
          type: array
          title: Nodes
      additionalProperties: false
      type: object
      required:
      - generated_at
      - nodes
      title: BudgetTree
...
    ConnectorStatus:
      properties:
        id:
          type: string
          title: Id
        type:
          type: string
          title: Type
        tier:
          $ref: '#/components/schemas/ConnectorTier'
        enabled:
          type: boolean
          title: Enabled
        kill_switch:
          type: boolean
          title: Kill Switch
        healthy:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Healthy
        latency_ms_p50:
          anyOf:
          - type: number
          - type: 'null'
          title: Latency Ms P50
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        spend_usd_day:
          type: number
          title: Spend Usd Day
          default: 0.0
      additionalProperties: false
      type: object
      required:
      - id
      - type
      - tier
      - enabled
      - kill_switch
      title: ConnectorStatus
...
    Elevation:
      properties:
        scope:
          type: string
          title: Scope
          examples:
          - tool:opencode.write
        until:
          type: string
          format: date-time
          title: Until
      additionalProperties: false
      type: object
      required:
      - scope
      - until
      title: Elevation
...
    FeedStatus:
      properties:
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        bundle_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Bundle Version
        issued_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Issued At
        loaded_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Loaded At
        entries:
          type: integer
          title: Entries
          default: 0
        verified:
          type: boolean
          title: Verified
          default: false
        last_sync_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Sync At
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
      additionalProperties: false
      type: object
      title: FeedStatus
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KillSwitchRequest:
      properties:
        engaged:
          type: boolean
          title: Engaged
        reason:
          type: string
          maxLength: 500
          minLength: 3
          title: Reason
      additionalProperties: false
      type: object
      required:
      - engaged
      - reason
      title: KillSwitchRequest
...
    McpServerInfo:
      properties:
        id:
          type: string
          title: Id
        transport:
          type: string
          enum:
          - streamable_http
          - stdio
          title: Transport
        origin:
          anyOf:
          - type: string
          - type: 'null'
          title: Origin
        allowed:
          type: boolean
          title: Allowed
        status:
          type: string
          enum:
          - ok
          - unreachable
          - blocked
          - unknown
          title: Status
          default: unknown
        tools_count:
          type: integer
          title: Tools Count
          default: 0
        protocol_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Protocol Version
        last_seen:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Seen
      additionalProperties: false
      type: object
      required:
      - id
      - transport
      - allowed
      title: McpServerInfo
    McpToolApprovalRequest:
      properties:
        reason:
          type: string
          maxLength: 500
          minLength: 3
          title: Reason
      additionalProperties: false
      type: object
      required:
      - reason
      title: McpToolApprovalRequest
    McpToolInfo:
      properties:
        id:
          type: string
          title: Id
          examples:
          - mail.send
        server:
          type: string
          title: Server
        name:
          type: string
          title: Name
        status:
          type: string
          enum:
          - pinned
          - pending_approval
          - drifted
          - quarantined
          title: Status
        pinned_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Pinned Hash
        current_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Hash
        labels:
          items:
            $ref: '#/components/schemas/ToolLabel'
          type: array
          title: Labels
        tier:
          anyOf:
          - $ref: '#/components/schemas/ToolTier'
          - type: 'null'
        first_seen:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: First Seen
        drift_detected_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Drift Detected At
        description_diff:
          anyOf:
          - type: string
          - type: 'null'
          title: Description Diff
      additionalProperties: false
      type: object
      required:
      - id
      - server
      - name
      - status
      title: McpToolInfo
    ModelInfo:
      properties:
        id:
          type: string
          title: Id
        connector:
          type: string
          title: Connector
        tier:
          $ref: '#/components/schemas/ConnectorTier'
        aliases:
          items:
            type: string
          type: array
          title: Aliases
        tags:
          additionalProperties:
            type: string
          type: object
          title: Tags
        data_classes:
          items:
            $ref: '#/components/schemas/DataClass'
          type: array
          title: Data Classes
        pricing:
          additionalProperties:
            type: number
          type: object
          title: Pricing
        artifact_status:
          type: string
          enum:
          - n/a
          - scanned_ok
          - scanned_bad
          - unscanned
          title: Artifact Status
          default: n/a
        enabled:
          type: boolean
          title: Enabled
          default: true
        available:
          type: boolean
          title: Available
          description: False when e.g. an env-referenced upstream name is unset.
          default: true
      additionalProperties: false
      type: object
      required:
      - id
      - connector
      - tier
      title: ModelInfo
...
    PublishSkillRequest:
      properties:
        skill_id:
          type: string
          pattern: ^skill/[a-z0-9][a-z0-9-]*$
          title: Skill Id
        model:
          type: string
          title: Model
        preset:
          $ref: '#/components/schemas/Preset'
          default: strict
        groups:
          items:
            type: string
          type: array
          minItems: 1
          title: Groups
        reason:
          type: string
          minLength: 3
          title: Reason
      additionalProperties: false
      type: object
      required:
      - skill_id
      - model
      - groups
      - reason
      title: PublishSkillRequest
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: Keycloak access token (JWT) or personal API key (`acl_...`).
      scheme: bearer
servers:
- url: http://localhost:8000